move telemetry.send into module
